Expanding the ranges of threshold parameters (TRG)

+++ src/trigger.h
@@ -14,7 +14,7 @@
 typedef struct __attribute__((packed)) _trigger_flg_t {
 	uint8_t 	trg_input	:	1; // GPIO_TRG pin input value (real level)
 	uint8_t 	trg_output	:	1; // GPIO_TRG pin output value (pull Up/Down)
-	uint8_t 	trigger_on	:	1; // Output GPIO_TRG pin is controlled according to the set parameters
+	uint8_t 	trigger_on	:	1; // Output GPIO_TRG pin is controlled according to the set parameters threshold temperature or humidity
 	uint8_t 	temp_out_on :	1; // Temperature trigger event
 	uint8_t 	humi_out_on :	1; // Humidity trigger event
 }trigger_flg_t;
@@ -22,8 +22,8 @@ typedef struct __attribute__((packed)) _trigger_flg_t {
 typedef struct __attribute__((packed)) _trigger_t {
 	int16_t temp_threshold; // x0.01°, Set temp threshold
 	int16_t humi_threshold; // x0.01%, Set humi threshold
-	int8_t temp_hysteresis; // Set temp hysteresis, -12,5 - +12,5 °C (-125..125)
-	int8_t humi_hysteresis; // Set humi hysteresis, -12,5 - +12,5 % (-125..125)
+	int16_t temp_hysteresis; // Set temp hysteresis, -327.67..327.67 °
+	int16_t humi_hysteresis; // Set humi hysteresis, -327.67..327.67 %
 	trigger_flg_t flg;
 }trigger_t;
 #define FEEP_SAVE_SIZE_TRG (sizeof(trg)-1)
